requirements.txt
vaderSentiment
tweepy
pandas
matplotlib
json
python-decouple